fix: handlerQA remove from reducer
